agentic-rag/graph/chains/generation.py
from langchain import hub
from langchain_core.output_parsers import StrOutputParser
from langchain_google_genai import ChatGoogleGenerativeAI


llm = ChatGoogleGenerativeAI(temperature=0.1, model="gemini-2.0-flash")
prompt = hub.pull("rlm/rag-prompt")

generation_chain = prompt | llm | StrOutputParser()
