agentic-rag/pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "black>=25.1.0",
    "grandalf>=0.8",
    "isort>=6.0.1",
    "langchain>=0.3.24",
    "langchain-chroma>=0.2.3",
    "langchain-community>=0.3.23",
    "langchain-google-genai>=2.1.4",
    "langchain-groq>=0.3.2",
    "langchain-ollama>=0.3.2",
    "langchain-tavily>=0.1.6",
    "langchainhub>=0.1.21",
    "langgraph>=0.4.0",
    "langsmith>=0.3.38",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "tiktoken>=0.9.0",
]
